Added global mapperConfig
