+++ business-modules/profiler-service/profiler-core/src/main/java/raff/stein/profiler/controller/mapper/UserPermissionDTOMapper.java
@@ -7,13 +7,13 @@ import raff.stein.profiler.model.UserPermission;
 import raff.stein.profiler.model.entity.mapper.PermissionEntityToPermissionMapper;
 import raff.stein.profiler.model.mapper.FeatureDTOToFeatureMapper;
 import raff.stein.profiler.model.mapper.SectionDTOToSectionMapper;
-import raff.stein.profiler.model.mapper.common.CommonMapperConfiguration;
+import raff.stein.profiler.model.mapper.common.ProfilerCommonMapperConfiguration;
 
 @Mapper(uses = {
         FeatureDTOToFeatureMapper.class,
         SectionDTOToSectionMapper.class,
         PermissionEntityToPermissionMapper.class},
-        config = CommonMapperConfiguration.class)
+        config = ProfilerCommonMapperConfiguration.class)
 public interface UserPermissionDTOMapper {
 
     UserPermissionDTOMapper MAPPER = Mappers.getMapper(UserPermissionDTOMapper.class);
