+++ business-modules/profiler-service/profiler-core/src/main/java/raff/stein/profiler/model/entity/mapper/FeatureEntityToFeatureMapper.java
@@ -4,9 +4,9 @@ import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
 import raff.stein.profiler.model.Feature;
 import raff.stein.profiler.model.entity.FeatureEntity;
-import raff.stein.profiler.model.mapper.common.CommonMapperConfiguration;
+import raff.stein.profiler.model.mapper.common.ProfilerCommonMapperConfiguration;
 
-@Mapper(config = CommonMapperConfiguration.class)
+@Mapper(config = ProfilerCommonMapperConfiguration.class)
 public interface FeatureEntityToFeatureMapper {
 
     FeatureEntityToFeatureMapper MAPPER = Mappers.getMapper(FeatureEntityToFeatureMapper.class);
