+++ business-modules/profiler-service/profiler-core/src/main/java/raff/stein/profiler/model/entity/mapper/PermissionEntityToPermissionMapper.java
@@ -4,9 +4,9 @@ import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
 import raff.stein.profiler.model.Permission;
 import raff.stein.profiler.model.entity.PermissionEntity;
-import raff.stein.profiler.model.mapper.common.CommonMapperConfiguration;
+import raff.stein.profiler.model.mapper.common.ProfilerCommonMapperConfiguration;
 
-@Mapper(config = CommonMapperConfiguration.class)
+@Mapper(config = ProfilerCommonMapperConfiguration.class)
 public interface PermissionEntityToPermissionMapper {
 
     PermissionEntityToPermissionMapper MAPPER = Mappers.getMapper(PermissionEntityToPermissionMapper.class);
