+++ business-modules/profiler-service/profiler-core/src/main/java/raff/stein/profiler/model/entity/mapper/SectionEntityToSectionMapper.java
@@ -4,9 +4,9 @@ import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
 import raff.stein.profiler.model.Section;
 import raff.stein.profiler.model.entity.SectionEntity;
-import raff.stein.profiler.model.mapper.common.CommonMapperConfiguration;
+import raff.stein.profiler.model.mapper.common.ProfilerCommonMapperConfiguration;
 
-@Mapper(config = CommonMapperConfiguration.class)
+@Mapper(config = ProfilerCommonMapperConfiguration.class)
 public interface SectionEntityToSectionMapper {
 
     SectionEntityToSectionMapper MAPPER = Mappers.getMapper(SectionEntityToSectionMapper.class);
