+++ business-modules/profiler-service/profiler-core/src/main/java/raff/stein/profiler/model/mapper/PermissionDTOToPermissionMapper.java
@@ -4,9 +4,9 @@ import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
 import org.openapitools.model.PermissionDTO;
 import raff.stein.profiler.model.Permission;
-import raff.stein.profiler.model.mapper.common.CommonMapperConfiguration;
+import raff.stein.profiler.model.mapper.common.ProfilerCommonMapperConfiguration;
 
-@Mapper(config = CommonMapperConfiguration.class)
+@Mapper(config = ProfilerCommonMapperConfiguration.class)
 public interface PermissionDTOToPermissionMapper {
 
     PermissionDTOToPermissionMapper MAPPER = Mappers.getMapper(PermissionDTOToPermissionMapper.class);
