--- business-modules/profiler-service/profiler-core/src/main/java/raff/stein/profiler/model/mapper/common/CommonMapperConfiguration.java
@@ -1,13 +1,0 @@
-package raff.stein.profiler.model.mapper.common;
-
-import org.mapstruct.InjectionStrategy;
-import org.mapstruct.MapperConfig;
-import org.mapstruct.ReportingPolicy;
-
-@MapperConfig(
-    unmappedTargetPolicy = ReportingPolicy.IGNORE,
-    unmappedSourcePolicy = ReportingPolicy.WARN,
-    injectionStrategy = InjectionStrategy.CONSTRUCTOR
-)
-public interface CommonMapperConfiguration {
-}
