+++ business-modules/profiler-service/profiler-core/src/main/java/raff/stein/profiler/model/mapper/common/ProfilerCommonMapperConfiguration.java
@@ -1,0 +1,8 @@
+package raff.stein.profiler.model.mapper.common;
+
+import org.mapstruct.MapperConfig;
+import raff.stein.platformcore.model.mapper.configuration.CommonMapperConfiguration;
+
+@MapperConfig(uses = CommonMapperConfiguration.class)
+public interface ProfilerCommonMapperConfiguration {
+}
