+++ business-modules/user-service/user-core/pom.xml
@@ -52,6 +52,10 @@
             <artifactId>jakarta.persistence-api</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.mapstruct</groupId>
+            <artifactId>mapstruct</artifactId>
+        </dependency>
+        <dependency>
             <groupId>raff.stein</groupId>
             <artifactId>user-api-data</artifactId>
             <version>1.0-SNAPSHOT</version>
