+++ business-modules/user-service/user-core/src/main/java/raff/stein/user/controller/mapper/UserDTOMapper.java
@@ -1,0 +1,8 @@
+package raff.stein.user.controller.mapper;
+
+import org.mapstruct.Mapper;
+import raff.stein.platformcore.model.mapper.configuration.CommonMapperConfiguration;
+
+@Mapper(config = CommonMapperConfiguration.class)
+public interface UserDTOMapper {
+}
