+++ core/platform-core/src/main/java/raff/stein/platformcore/model/mapper/configuration/CommonMapperConfiguration.java
@@ -1,0 +1,13 @@
+package raff.stein.platformcore.model.mapper.configuration;
+
+import org.mapstruct.InjectionStrategy;
+import org.mapstruct.MapperConfig;
+import org.mapstruct.ReportingPolicy;
+
+@MapperConfig(
+        unmappedTargetPolicy = ReportingPolicy.IGNORE,
+        unmappedSourcePolicy = ReportingPolicy.WARN,
+        injectionStrategy = InjectionStrategy.CONSTRUCTOR
+)
+public interface CommonMapperConfiguration {
+}
